refactor(portal): organizar src por feature em vez de por camada
O src/ tinha quatro convenções ao mesmo tempo: pasta técnica (Livewire/,
DTOs/), pasta de feature (Articles/, Retrospective/, ShortLink/) e classe
solta na raiz (SitemapController). O caso dos artigos era o mais visível —
ArticlesPage em Livewire/ e o read model que ele injeta em Articles/, dois
ramos distantes da árvore para uma coisa só.

Agora toda pasta de primeiro nível é uma feature e reúne a entrada com o
que ela usa:

  Articles/      ArticlesPage + ArticleFeed + Article + VOs
  Home/          Homepage + HeroSection
  Retrospective/ CommunityRetrospectivePage + CommunityRetrospective + Filters
  ShortLink/     ShortLinkRedirectController  (já estava assim)
  Sitemap/       SitemapController
  SocialLinks/   SocialLinksPage + SocialLink

Livewire/ e DTOs/ deixam de existir.

Só arquivos movidos e namespaces trocados; nenhum comportamento muda. As
rotas full-page referenciam a classe direto e <livewire:hero-section /> tem
alias explícito no provider, então nada depende de convenção de namespace.
