+++ app-modules/portal/tests/Feature/ArticlesPageTest.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 use He4rt\Portal\Articles\Article;
 use He4rt\Portal\Articles\ArticleFeed;
-use He4rt\Portal\Livewire\ArticlesPage;
+use He4rt\Portal\Articles\ArticlesPage;
 use Illuminate\Support\Facades\Cache;
 use Illuminate\Support\Facades\Http;
